+++ threshold_resource_estimator.hpp
@@ -2,12 +2,19 @@
 
 #include <mesos/module/resource_estimator.hpp>
 
+#include <functional>
+
+#include <stout/bytes.hpp>
+
 namespace os {
     struct Load;
-    struct Memory;
 }
 
 namespace com { namespace blue_yonder {
+
+namespace os {
+    struct MemInfo;
+}
 
 class ThresholdResourceEstimatorProcess;
 
@@ -15,8 +22,8 @@ class ThresholdResourceEstimator : public mesos::slave::ResourceEstimator
 {
 public:
     ThresholdResourceEstimator(
-        std::function<Try<os::Load>()> const & load,
-        std::function<Try<os::Memory>()> const & memory,
+        std::function<Try<::os::Load>()> const & load,
+        std::function<Try<os::MemInfo>()> const & memory,
         mesos::Resources const & fixed,
         Option<double> const & loadThreshold1Min,
         Option<double> const & loadThreshold5Min,
@@ -29,8 +36,8 @@ public:
 
 private:
     process::Owned<ThresholdResourceEstimatorProcess> process;
-    std::function<Try<os::Load>()> const load;
-    std::function<Try<os::Memory>()> const memory;
+    std::function<Try<::os::Load>()> const load;
+    std::function<Try<os::MemInfo>()> const memory;
     mesos::Resources const fixed;
     Option<double> const loadThreshold1Min;
     Option<double> const loadThreshold5Min;